fixed: compile warning.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@10764 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/guilib/GUITextBox.cpp
@@ -543,7 +543,7 @@ void CGUITextBox::SetAutoScrolling(const TiXmlNode *node)
     scroll->Attribute("time", &m_autoScrollTime);
     if (scroll->FirstChild())
       m_autoScrollCondition = g_infoManager.TranslateString(scroll->FirstChild()->ValueStr());
-    float repeatTime;
+    int repeatTime;
     if (scroll->Attribute("repeat", &repeatTime))
       m_autoScrollRepeatAnim = CAnimation::CreateFader(100, 0, repeatTime, 1000);
   }
